+++ src/main/java/br/com/feliperochasi/AdopetConsoleApplication.java
@@ -1,5 +1,6 @@
 package br.com.feliperochasi;
 
+import br.com.feliperochasi.service.AbrigoService;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
@@ -17,6 +18,7 @@ import java.util.Scanner;
 public class AdopetConsoleApplication {
 
     public static void main(String[] args) {
+        AbrigoService abrigoService = new AbrigoService();
         System.out.println("##### BOAS VINDAS AO SISTEMA ADOPET CONSOLE #####");
         try {
             int opcaoEscolhida = 0;
@@ -32,9 +34,9 @@ public class AdopetConsoleApplication {
                 opcaoEscolhida = Integer.parseInt(textoDigitado);
 
                 if (opcaoEscolhida == 1) {
-                    listarAbrigo();
+                    abrigoService.listarAbrigo();
                 } else if (opcaoEscolhida == 2) {
-                    cadastrarAbrigo();
+                    abrigoService.cadastrarAbrigo();
                 } else if (opcaoEscolhida == 3) {
                     listarPetsDoAbrigo();
                 } else if (opcaoEscolhida == 4) {
@@ -49,48 +51,6 @@ public class AdopetConsoleApplication {
             System.out.println("Finalizando o programa...");
         } catch (Exception e) {
             e.printStackTrace();
-        }
-    }
-
-    private static void listarAbrigo() throws IOException, InterruptedException {
-        HttpClient client = HttpClient.newHttpClient();
-        String uri = "http://localhost:8080/abrigos";
-        HttpResponse<String> response = dispararRequiscaoGet(client, uri);
-        String responseBody = response.body();
-        JsonArray jsonArray = JsonParser.parseString(responseBody).getAsJsonArray();
-        System.out.println("Abrigos cadastrados:");
-        for (JsonElement element : jsonArray) {
-            JsonObject jsonObject = element.getAsJsonObject();
-            long id = jsonObject.get("id").getAsLong();
-            String nome = jsonObject.get("nome").getAsString();
-            System.out.println(id +" - " +nome);
-        }
-    }
-
-    private static void cadastrarAbrigo() throws IOException, InterruptedException {
-        System.out.println("Digite o nome do abrigo:");
-        String nome = new Scanner(System.in).nextLine();
-        System.out.println("Digite o telefone do abrigo:");
-        String telefone = new Scanner(System.in).nextLine();
-        System.out.println("Digite o email do abrigo:");
-        String email = new Scanner(System.in).nextLine();
-
-        JsonObject json = new JsonObject();
-        json.addProperty("nome", nome);
-        json.addProperty("telefone", telefone);
-        json.addProperty("email", email);
-
-        HttpClient client = HttpClient.newHttpClient();
-        String uri = "http://localhost:8080/abrigos";
-        var response = dispararRequisicaoPost(client, uri, json);
-        int statusCode = response.statusCode();
-        String responseBody = response.body();
-        if (statusCode == 200) {
-            System.out.println("Abrigo cadastrado com sucesso!");
-            System.out.println(responseBody);
-        } else if (statusCode == 400 || statusCode == 500) {
-            System.out.println("Erro ao cadastrar o abrigo:");
-            System.out.println(responseBody);
         }
     }
