+++ src/main/java/br/com/feliperochasi/service/PetService.java
@@ -1,0 +1,4 @@
+package br.com.feliperochasi.service;
+
+public class PetService {
+}
